move persistence to middleware
