--- app/persistence/index.js
@@ -1,3 +1,0 @@
-import persistence from './persistence'
-
-export default persistence
