+++ app/state/store/store.js
@@ -5,10 +5,9 @@ import websocket from 'hg/state/middlewares/websocket'
 import logger from 'hg/state/middlewares/logger'
 import navigation from 'hg/state/middlewares/navigation'
 import persistence from 'hg/state/middlewares/persistence'
-import { default as _persistence } from 'hg/persistence'
 
 const middlewares = [
-  persistence(_persistence),
+  persistence,
   websocket,
   navigation(NavigationActions),
   logger
